requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
deprecation==2.1.0
exceptiongroup==1.2.2
fake-useragent==2.0.3
frozenlist==1.5.0
gotrue==2.11.2
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
packaging==24.2
postgrest==0.19.3
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.2.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
storage3==0.11.1
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.2
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
yarl==1.18.3
